+++ libs/application/templates/social-insurance-administration/core/project.json
@@ -22,6 +22,12 @@
         "jestConfig": "libs/application/templates/social-insurance-administration/core/jest.config.js",
         "passWithNoTests": true
       }
+    },
+    "extract-strings": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "yarn ts-node -P libs/localization/tsconfig.lib.json libs/localization/scripts/extract 'libs/application/templates/social-insurance-administration/core/src/messages.ts'"
+      }
     }
   },
   "tags": ["scope:application-system", "lib:application-system"]
