+++ libs/application/templates/social-insurance-administration/household-supplement/src/lib/HouseholdSupplementTemplate.ts
@@ -64,6 +64,7 @@ const HouseholdSupplementTemplate: ApplicationTemplate<
   featureFlag: Features.householdSupplementApplication,
   translationNamespaces: [
     ApplicationConfigurations.HouseholdSupplement.translation,
+    'sia.application',
   ],
   dataSchema,
   allowMultipleApplicationsInDraft: false,
