+++ libs/application/templates/social-insurance-administration/old-age-pension/src/fields/Review/review-groups/Period.tsx
@@ -1,7 +1,8 @@
 import { DataValue, ReviewGroup } from '@island.is/application/ui-components'
 import { GridColumn, GridRow } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
-import { errorMessages, oldAgePensionFormMessage } from '../../../lib/messages'
+import { oldAgePensionFormMessage } from '../../../lib/messages'
+import { errorMessages } from '@island.is/application/templates/social-insurance-administration-core/messages'
 import { ReviewGroupProps } from './props'
 import { MONTHS } from '../../../lib/constants'
 import {
