+++ libs/clients/social-insurance-administration/src/clientConfig.yaml
@@ -253,7 +253,6 @@ components:
             - selfEmployed
             - rentalAgreement
             - schoolConfirmation
-            - undefined
         file:
           type: string
     Response:
